Improve harmony build
