+++ packages/harmony/package.json
@@ -4,9 +4,9 @@
   "description": "The Audius Design System",
   "license": "ISC",
   "author": "Audius",
-  "main": "dist/index.es.js",
-  "module": "dist/index.es.js",
-  "jsnext:main": "dist/index.es.js",
+  "main": "dist/index.js",
+  "module": "dist/index.js",
+  "jsnext:main": "dist/index.js",
   "types": "dist/index.d.ts",
   "files": [
     "dist"
@@ -64,7 +64,6 @@
     "eslint": "8.19.0",
     "eslint-plugin-storybook": "0.6.13",
     "lodash": "4.17.21",
-    "postcss-custom-properties": "9.2.0",
     "prettier": "2.7.1",
     "react": "18.2.0",
     "react-docgen-typescript-loader": "3.7.2",
@@ -74,14 +73,10 @@
     "react-perfect-scrollbar": "1.5.8",
     "react-use": "15.3.8",
     "react-use-measure": "2.1.1",
-    "rollup": "1.32.1",
-    "rollup-plugin-babel": "4.4.0",
-    "rollup-plugin-copy": "3.5.0",
+    "rollup": "4.1.4",
     "rollup-plugin-json": "4.0.0",
-    "rollup-plugin-node-resolve": "5.2.0",
     "rollup-plugin-postcss": "4.0.2",
-    "rollup-plugin-typescript2": "0.24.3",
-    "rollup-plugin-url": "2.2.4",
+    "rollup-plugin-typescript2": "0.36.0",
     "storybook": "7.4.2",
     "storybook-addon-smart-knobs": "6.0.2",
     "tsconfig-paths-webpack-plugin": "3.5.2",
