+++ packages/harmony/src/assets/styles/harmony.css
@@ -1,0 +1,2 @@
+.test {
+}
